Redefine DataKey contracts with interfaces instead of Javadoc - 1st round of fixes
