--- frontendActions/src/main/java/com/virtuslab/gitmachete/frontend/actions/common/IExpectsKeyBranchLayout.java
@@ -1,12 +1,0 @@
-package com.virtuslab.gitmachete.frontend.actions.common;
-
-import com.intellij.openapi.actionSystem.AnActionEvent;
-import io.vavr.control.Option;
-
-import com.virtuslab.branchlayout.api.IBranchLayout;
-
-public interface IExpectsKeyBranchLayout extends IExpectsKeyGitMacheteRepository {
-  default Option<IBranchLayout> getBranchLayout(AnActionEvent anActionEvent) {
-    return getGitMacheteRepository(anActionEvent).flatMap(repository -> repository.getBranchLayout());
-  }
-}
