--- frontendActions/src/main/java/com/virtuslab/gitmachete/frontend/actions/common/IExpectsKeyGitMacheteRepository.java
@@ -1,26 +1,0 @@
-package com.virtuslab.gitmachete.frontend.actions.common;
-
-import com.intellij.openapi.actionSystem.AnActionEvent;
-import io.vavr.control.Option;
-
-import com.virtuslab.gitmachete.backend.api.BaseGitMacheteBranch;
-import com.virtuslab.gitmachete.backend.api.IGitMacheteRepository;
-import com.virtuslab.gitmachete.frontend.datakeys.DataKeys;
-
-public interface IExpectsKeyGitMacheteRepository {
-  default Option<IGitMacheteRepository> getGitMacheteRepository(AnActionEvent anActionEvent) {
-    return Option.of(anActionEvent.getData(DataKeys.KEY_GIT_MACHETE_REPOSITORY));
-  }
-
-  default Option<String> getCurrentBranchNameIfManaged(AnActionEvent anActionEvent) {
-    return getCurrentMacheteBranchIfManaged(anActionEvent).map(branch -> branch.getName());
-  }
-
-  default Option<BaseGitMacheteBranch> getCurrentMacheteBranchIfManaged(AnActionEvent anActionEvent) {
-    return getGitMacheteRepository(anActionEvent).flatMap(repository -> repository.getCurrentBranchIfManaged());
-  }
-
-  default Option<BaseGitMacheteBranch> getGitMacheteBranchByName(AnActionEvent anActionEvent, String branchName) {
-    return getGitMacheteRepository(anActionEvent).flatMap(r -> r.getBranchByName(branchName));
-  }
-}
