+++ frontendActions/src/main/java/com/virtuslab/gitmachete/frontend/actions/contextmenu/FastForwardParentToMatchSelectedBranchAction.java
@@ -3,8 +3,8 @@ package com.virtuslab.gitmachete.frontend.actions.contextmenu;
 import com.intellij.openapi.actionSystem.AnActionEvent;
 import io.vavr.control.Option;
 
-import com.virtuslab.gitmachete.frontend.actions.common.BaseFastForwardParentToMatchBranchAction;
-import com.virtuslab.gitmachete.frontend.actions.common.IExpectsKeySelectedBranchName;
+import com.virtuslab.gitmachete.frontend.actions.base.BaseFastForwardParentToMatchBranchAction;
+import com.virtuslab.gitmachete.frontend.actions.expectedkeys.IExpectsKeySelectedBranchName;
 
 public class FastForwardParentToMatchSelectedBranchAction extends BaseFastForwardParentToMatchBranchAction
     implements
