+++ frontendActions/src/main/java/com/virtuslab/gitmachete/frontend/actions/contextmenu/PullSelectedBranchAction.java
@@ -3,8 +3,8 @@ package com.virtuslab.gitmachete.frontend.actions.contextmenu;
 import com.intellij.openapi.actionSystem.AnActionEvent;
 import io.vavr.control.Option;
 
-import com.virtuslab.gitmachete.frontend.actions.common.BasePullBranchAction;
-import com.virtuslab.gitmachete.frontend.actions.common.IExpectsKeySelectedBranchName;
+import com.virtuslab.gitmachete.frontend.actions.base.BasePullBranchAction;
+import com.virtuslab.gitmachete.frontend.actions.expectedkeys.IExpectsKeySelectedBranchName;
 
 public class PullSelectedBranchAction extends BasePullBranchAction implements IExpectsKeySelectedBranchName {
   @Override
