+++ frontendActions/src/main/java/com/virtuslab/gitmachete/frontend/actions/contextmenu/PushSelectedBranchAction.java
@@ -3,8 +3,8 @@ package com.virtuslab.gitmachete.frontend.actions.contextmenu;
 import com.intellij.openapi.actionSystem.AnActionEvent;
 import io.vavr.control.Option;
 
-import com.virtuslab.gitmachete.frontend.actions.common.BasePushBranchAction;
-import com.virtuslab.gitmachete.frontend.actions.common.IExpectsKeySelectedBranchName;
+import com.virtuslab.gitmachete.frontend.actions.base.BasePushBranchAction;
+import com.virtuslab.gitmachete.frontend.actions.expectedkeys.IExpectsKeySelectedBranchName;
 
 public class PushSelectedBranchAction extends BasePushBranchAction implements IExpectsKeySelectedBranchName {
   @Override
