+++ frontendActions/src/main/java/com/virtuslab/gitmachete/frontend/actions/contextmenu/ResetSelectedBranchToRemoteAction.java
@@ -4,7 +4,7 @@ import com.intellij.openapi.actionSystem.AnActionEvent;
 import io.vavr.control.Option;
 
 import com.virtuslab.gitmachete.frontend.actions.common.BaseResetBranchToRemoteAction;
-import com.virtuslab.gitmachete.frontend.actions.common.IExpectsKeySelectedBranchName;
+import com.virtuslab.gitmachete.frontend.actions.expectedkeys.IExpectsKeySelectedBranchName;
 
 public class ResetSelectedBranchToRemoteAction extends BaseResetBranchToRemoteAction implements IExpectsKeySelectedBranchName {
   @Override
