+++ frontendActions/src/main/java/com/virtuslab/gitmachete/frontend/actions/contextmenu/SlideOutSelectedBranchAction.java
@@ -3,8 +3,8 @@ package com.virtuslab.gitmachete.frontend.actions.contextmenu;
 import com.intellij.openapi.actionSystem.AnActionEvent;
 import io.vavr.control.Option;
 
-import com.virtuslab.gitmachete.frontend.actions.common.BaseSlideOutBranchAction;
-import com.virtuslab.gitmachete.frontend.actions.common.IExpectsKeySelectedBranchName;
+import com.virtuslab.gitmachete.frontend.actions.base.BaseSlideOutBranchAction;
+import com.virtuslab.gitmachete.frontend.actions.expectedkeys.IExpectsKeySelectedBranchName;
 
 public class SlideOutSelectedBranchAction extends BaseSlideOutBranchAction implements IExpectsKeySelectedBranchName {
   @Override
