+++ frontendActions/src/main/java/com/virtuslab/gitmachete/frontend/actions/expectedkeys/IExpectsKeyGitMacheteRepository.java
@@ -1,0 +1,31 @@
+package com.virtuslab.gitmachete.frontend.actions.expectedkeys;
+
+import com.intellij.openapi.actionSystem.AnActionEvent;
+import io.vavr.control.Option;
+
+import com.virtuslab.branchlayout.api.IBranchLayout;
+import com.virtuslab.gitmachete.backend.api.IGitMacheteBranch;
+import com.virtuslab.gitmachete.backend.api.IGitMacheteRepository;
+import com.virtuslab.gitmachete.frontend.datakeys.DataKeys;
+
+public interface IExpectsKeyGitMacheteRepository {
+  default Option<IGitMacheteRepository> getGitMacheteRepository(AnActionEvent anActionEvent) {
+    return Option.of(anActionEvent.getData(DataKeys.KEY_GIT_MACHETE_REPOSITORY));
+  }
+
+  default Option<IBranchLayout> getBranchLayout(AnActionEvent anActionEvent) {
+    return getGitMacheteRepository(anActionEvent).flatMap(repository -> repository.getBranchLayout());
+  }
+
+  default Option<IGitMacheteBranch> getCurrentMacheteBranchIfManaged(AnActionEvent anActionEvent) {
+    return getGitMacheteRepository(anActionEvent).flatMap(repository -> repository.getCurrentBranchIfManaged());
+  }
+
+  default Option<String> getCurrentBranchNameIfManaged(AnActionEvent anActionEvent) {
+    return getCurrentMacheteBranchIfManaged(anActionEvent).map(branch -> branch.getName());
+  }
+
+  default Option<IGitMacheteBranch> getGitMacheteBranchByName(AnActionEvent anActionEvent, String branchName) {
+    return getGitMacheteRepository(anActionEvent).flatMap(r -> r.getBranchByName(branchName));
+  }
+}
