+++ frontendActions/src/main/java/com/virtuslab/gitmachete/frontend/actions/toolbar/PullCurrentBranchAction.java
@@ -3,7 +3,7 @@ package com.virtuslab.gitmachete.frontend.actions.toolbar;
 import com.intellij.openapi.actionSystem.AnActionEvent;
 import io.vavr.control.Option;
 
-import com.virtuslab.gitmachete.frontend.actions.common.BasePullBranchAction;
+import com.virtuslab.gitmachete.frontend.actions.base.BasePullBranchAction;
 
 public class PullCurrentBranchAction extends BasePullBranchAction {
   @Override
