+++ frontendActions/src/main/java/com/virtuslab/gitmachete/frontend/actions/toolbar/RebaseCurrentBranchOntoParentAction.java
@@ -3,7 +3,7 @@ package com.virtuslab.gitmachete.frontend.actions.toolbar;
 import com.intellij.openapi.actionSystem.AnActionEvent;
 import io.vavr.control.Option;
 
-import com.virtuslab.gitmachete.frontend.actions.common.BaseRebaseBranchOntoParentAction;
+import com.virtuslab.gitmachete.frontend.actions.base.BaseRebaseBranchOntoParentAction;
 
 public class RebaseCurrentBranchOntoParentAction extends BaseRebaseBranchOntoParentAction {
   @Override
